+++ src/main/java/cart/controller/cart/CartViewController.java
@@ -1,0 +1,14 @@
+package cart.controller.cart;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class CartViewController {
+
+
+    @GetMapping("/cart")
+    public String getCartPage() {
+        return "cart";
+    }
+}
