+++ src/main/java/cart/controller/setting/SettingViewController.java
@@ -1,0 +1,13 @@
+package cart.controller.setting;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class SettingViewController {
+
+    @GetMapping("/settings")
+    public String getSettingsPage() {
+        return "settings";
+    }
+}
